bugfix: couldn't attach the content of second chunk of spaces in case two chunks in a row contain only spaces
While attaching the content of space chunks to the previous chunk, if the previous chunk was a space chunk and its content already attached, the only thing remaining is a bool False as space-holder.
In this case, if the current chunk is also a space chunk, it couldn't attach to the previous chunk.

Now goes back to the last valid chunk and attaches content to it.